Commit first and second version
This is the first version of the code let you control the intensity of two groups of LEDs independently, and one group would always be connected to the potentiometer. The second version has all LEDs at the same intensity, and toggle between a programming mode where the intensity depends on the position of the potentiometer, and another where the intensity is not affected by the potentiometer.